Move the modified _mirror_groups() method to a separate module to retain license
